feat: move exceptions to errors package
