+++ qubership-nifi-tools/qubership-nifi-flow-diff-core/src/main/java/org/qubership/nifi/flowdiff/flow/ComponentIndex.java
@@ -1,6 +1,7 @@
 package org.qubership.nifi.flowdiff.flow;
 
 import com.fasterxml.jackson.databind.JsonNode;
+import org.qubership.nifi.flowdiff.error.FlowParseException;
 
 import java.util.ArrayList;
 import java.util.Collections;
