+++ qubership-nifi-tools/qubership-nifi-flow-diff-tool/src/main/java/org/qubership/nifi/maven/flowdiff/mojo/DiffMojo.java
@@ -7,7 +7,7 @@ import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.qubership.nifi.flowdiff.error.FlowDiffExecutionException;
 import org.qubership.nifi.flowdiff.error.FlowDiffInputException;
-import org.qubership.nifi.flowdiff.flow.FlowParseException;
+import org.qubership.nifi.flowdiff.error.FlowParseException;
 import org.qubership.nifi.flowdiff.report.ReportModel;
 import org.qubership.nifi.flowdiff.service.FlowDiffService;
